+++ shuup_tests/core/test_order_refund_utils.py
@@ -154,6 +154,211 @@ def test_order_refunds_with_multiple_suppliers():
 
 
 @pytest.mark.django_db
+def test_order_arbitrary_refunds_with_multiple_suppliers():
+    shop = get_default_shop()
+    supplier1 = Supplier.objects.create(identifier="1", name="supplier1")
+    supplier1.shops.add(shop)
+    supplier2 = Supplier.objects.create(identifier="2")
+    supplier2.shops.add(shop)
+    supplier3 = Supplier.objects.create(identifier="3", name="s")
+    supplier3.shops.add(shop)
+
+    product1 = create_product("sku1", shop=shop, default_price=10)
+    shop_product1 = product1.get_shop_instance(shop=shop)
+    shop_product1.suppliers.set([supplier1, supplier2, supplier3])
+
+    product2 = create_product("sku2", shop=shop, default_price=10)
+    shop_product2 = product1.get_shop_instance(shop=shop)
+    shop_product2.suppliers.set([supplier1, supplier2])
+
+    product3 = create_product("sku3", shop=shop, default_price=10, shipping_mode=ShippingMode.NOT_SHIPPED)
+    shop_product3 = product1.get_shop_instance(shop=shop)
+    shop_product3.suppliers.set([supplier3])
+
+    product_quantities = {
+        supplier1: {
+            product1: 5,
+            product2: 6
+        },
+        supplier2: {
+            product1: 3,
+            product2: 13
+        },
+        supplier3: {
+            product1: 1,
+            product3: 50
+        }
+    }
+
+    def get_quantity(supplier, product):
+        return product_quantities[supplier.pk][product.pk]
+
+    order = create_empty_order(shop=shop)
+    order.full_clean()
+    order.save()
+
+    for supplier, product_data in six.iteritems(product_quantities):
+        for product, quantity in six.iteritems(product_data):
+            add_product_to_order(order, supplier, product, quantity, 5)
+
+    # Lines without quantity shouldn't affect refunds
+    other_line = OrderLine(
+        order=order, type=OrderLineType.OTHER, text="This random line for textual information", quantity=0)
+    other_line.save()
+    order.lines.add(other_line)
+
+    order.cache_prices()
+    order.create_payment(order.taxful_total_price)
+    assert order.is_paid()
+
+    # All supplier should be able to refund the order
+    assert order.can_create_refund()
+    assert order.can_create_refund(supplier1)
+    assert order.can_create_refund(supplier2)
+    assert order.can_create_refund(supplier3)
+
+    # Step by step refund lines for supplier1
+    assert order.can_create_refund()
+    assert order.get_total_unrefunded_amount(supplier1).value == Decimal("55")  # 11 * 5
+    assert order.get_total_unrefunded_amount().value == Decimal("390")  # 55 + 80 + 255
+    proudct1_line_for_supplier1 = order.lines.filter(supplier=supplier1, product=product1).first()
+    supplier1_refund_data = [
+        {
+            "line": proudct1_line_for_supplier1,
+            "quantity": proudct1_line_for_supplier1.quantity,
+            "amount": order.shop.create_price(20).amount,  # Line total is 5 * 5 = 25
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier1_refund_data)
+    assert order.get_total_unrefunded_amount(supplier1).value == Decimal("35")
+
+    order.create_refund([{
+        "line": "amount",
+        "quantity": 1,
+        "amount": order.shop.create_price(30).amount
+    }], supplier=supplier1)
+
+    assert order.get_total_unrefunded_amount(supplier1).value == Decimal("5")
+
+    order.create_refund([{
+        "line": "amount",
+        "quantity": 1,
+        "amount": order.shop.create_price(5).amount
+    }], supplier=supplier1)
+
+    assert order.get_total_unrefunded_amount(supplier1).value == Decimal("0")
+    assert order.can_create_refund(supplier1)  # Some quantity still left to refund
+
+    proudct2_line_for_supplier1 = order.lines.filter(supplier=supplier1, product=product2).first()
+    supplier1_restock_refund_data = [
+        {
+            "line": proudct2_line_for_supplier1,
+            "quantity": proudct2_line_for_supplier1.quantity,
+            "amount": order.shop.create_price(0).amount,  # Line total is 5 * 5 = 25
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier1_restock_refund_data)
+    assert not order.can_create_refund(supplier1)
+
+    # Step by step refund lines for supplier2
+    assert order.can_create_refund()
+    assert order.get_total_unrefunded_amount(supplier2).value == Decimal("80")  # 16 * 5
+    assert order.get_total_unrefunded_amount().value == Decimal("335")  # 80 + 255
+    proudct2_line_for_supplier2 = order.lines.filter(supplier=supplier2, product=product2).first()
+    supplier2_refund_data = [
+        {
+            "line": proudct2_line_for_supplier2,
+            "quantity": 10,
+            "amount": order.shop.create_price(50).amount,  # Line total is 13 * 5 = 65
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier2_refund_data)
+    assert order.get_total_unrefunded_amount(supplier2).value == Decimal("30")
+
+    order.create_refund([{
+        "line": "amount",
+        "quantity": 1,
+        "amount": order.shop.create_price(5).amount
+    }], supplier=supplier2)
+
+    assert order.get_total_unrefunded_amount(supplier2).value == Decimal("25")
+
+    order.create_refund([{
+        "line": "amount",
+        "quantity": 1,
+        "amount": order.shop.create_price(25).amount
+    }], supplier=supplier2)
+
+    assert order.get_total_unrefunded_amount(supplier2).value == Decimal("0")
+    assert order.can_create_refund(supplier2)  # Some quantity still left to refund
+
+    supplier2_restock_refund_data = [
+        {
+            "line": proudct2_line_for_supplier2,
+            "quantity": 3,
+            "amount": order.shop.create_price(0).amount,  # Line total is 5 * 5 = 25
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier2_restock_refund_data)
+
+    proudct1_line_for_supplier2 = order.lines.filter(supplier=supplier2, product=product1).first()
+    supplier1_restock_refund_data = [
+        {
+            "line": proudct1_line_for_supplier2,
+            "quantity": proudct1_line_for_supplier2.quantity,
+            "amount": order.shop.create_price(0).amount,  # Line total is 5 * 5 = 25
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier1_restock_refund_data)
+    assert not order.can_create_refund(supplier2)
+
+    # Step by step refund lines for supplier3
+    assert order.can_create_refund()
+    assert order.get_total_unrefunded_amount(supplier3).value == Decimal("255")  # 51 * 5
+    assert order.get_total_unrefunded_amount().value == Decimal("255")  # 255
+
+    order.create_refund([{
+        "line": "amount",
+        "quantity": 1,
+        "amount": order.shop.create_price(55).amount
+    }], supplier=supplier3)
+
+    assert order.get_total_unrefunded_amount(supplier3).value == Decimal("200")
+
+    proudct3_line_for_supplier3 = order.lines.filter(supplier=supplier3, product=product3).first()
+    supplier3_refund_data = [
+        {
+            "line": proudct3_line_for_supplier3,
+            "quantity": 50,
+            "amount": order.shop.create_price(200).amount,  # Line total is 13 * 5 = 65
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier3_refund_data)
+    assert order.get_total_unrefunded_amount(supplier2).value == Decimal("0")
+    assert order.get_total_unrefunded_amount().value == Decimal("0")
+    assert order.can_create_refund(supplier3)  # Some quantity still left to refund
+
+    proudct1_line_for_supplier3 = order.lines.filter(supplier=supplier3, product=product1).first()
+    supplier3_restock_refund_data = [
+        {
+            "line": proudct1_line_for_supplier3,
+            "quantity": proudct1_line_for_supplier3.quantity,
+            "amount": order.shop.create_price(0).amount,  # Line total is 5 * 5 = 25
+            "restock_products": True
+        }
+    ]
+    order.create_refund(supplier3_restock_refund_data)
+    assert not order.can_create_refund(supplier3)
+    assert not order.can_create_refund()
+
+
+@pytest.mark.django_db
 def test_order_refunds_with_other_lines():
     shop = get_default_shop()
     supplier = Supplier.objects.create(identifier="1", name="supplier1")