+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Fixed
+
+- Core: include arbitrary refunds for max refundable amount
+
+
 ## [2.1.10] - 2020-09-29
 
 ### Fixed